agents/requirements.txt
uagents>=0.4.0
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0
asyncio-mqtt>=0.13.0
websockets>=11.0.0
aiohttp>=3.8.0
google-generativeai>=0.3.0
